немного изменил код в презентере
